various fixes

+++ app/controllers/likes_controller.rb
@@ -6,20 +6,22 @@ def create
      if like.save
      	flash[:notice] = "Bookmark favorited."
       redirect_to "/topics"
-       # Add code to generate a success flash and redirect to @bookmark
+       
      else
-       # Add code to generate a failure flash and redirect to @bookmark
+		flash[:notice] = "Bookmark could not favorited, please try again or contact support."
      end
  end
 
-def destroy
-     @bookmark = Bookmark.find(params[:bookmark_id])
-     like = current_user.likes.build(bookmark: @bookmark)
- 
-     if like.destroy
-       # Flash success and redirect to @bookmark
-     else
-       # Flash error and redirect to @bookmark
-     end
-end
+  def destroy
+    @bookmark = Bookmark.find(params[:bookmark_id])
+    like = current_user.likes.where(bookmark_id: @bookmark.id).first
+    if like
+      if like.destroy
+        flash[:notice] = "Bookmark un-favorited."
+        redirect_to "/topics"
+      else
+        flash[:notice] = "Bookmark could not un-favorited, please try again or contact support."
+      end
+    end
+  end
 end
